[Docs] Add agdq17-layouts to projects list
obs-websocket-js was used in `agdq17-layouts` to observe the active scene in OBS and change the behavior of the graphics accordingly. We also had a custom build of `obs-websocket` which could control our custom cropping plugin. Whenever we finished a run and hit "next run" on the schedule, `agdq17-layouts` would tell `obs-websocket` to reset the cropping.