test: test actual build output (#138)
Currently, tests are run against the un-transformed source found in `lib`. This is potentially hazardous, because it is possible for tests to pass but for the published `npm` package to be broken, because the published package is a transformed output of these sources files.

This change ensures that we are always testing against whatever the `main` entrypoint defined in `package.json` is, which helps avoid this scenario where tests pass but the build might be broken.

As a side effect of this, the order of the `script` steps in `.travis.yml` has been swapped, because we now can't run the tests until the build output is available.