chore: remove es5 build (#139)
It is doubtful that there are many users of obs-websocket-js still using environments which are limited to ES5. This is a layer of abstraction that I believe we can safely remove.

BREAKING CHANGE: The es5 build has been removed. Users limited to es5 environments should continue using the latest 2.x release.