test: fix ava config
The `test:ava` script was missing an extra `--`, meaning that the `--serial --timeout=10s` wasn't actually doing anything.

Instead of fixing this by adding that `--`, I figured it was best to update `ava` and leverage its expanded ability to be configured via a stanza in `package.json`. We already have this stanza, so making more use of it just made sense to me, rather than spreading our configuration between scripts and the stanza.