test: remove useless await
Constructors cannot be async, so putting an `await` in front of `new WebSocket.Server` does nothing. So, I removed it.

Since this was the only `await` present in `makeServer()`, I was able to remove the `async` from this method entirely. It does not seem to negatively impact the tests at all.